Start with a 20KB space for the incoming json
This will prevent a whole whack of realloc-ing with every
curl-writefunc call.  I chose 20KB since the typical JSON size is
12KB.  I figure this is enough for the usual issues.  For issues that
are bigger, the array grows by a scaling factor of 2 * new-data.

It's not a big deal either way, since this is a low-computation
program.  I just like looking at the actual data that we are loading
and working with that.  This number may change in the future,
depending on JIRA and my personal usage.