We shouldn't need a custom string container
This change has us using a simple C-string instead of our
growable-string struct.  It's not necessary to do that.  Just use
strlen when necessary.  We shouldn't use non-standard datastructures
till the standard ones are obviously insufficient.

Just send a pointer to the string to Curl, and it'll append to that
string on it's own, and grow it as necessary.